Move workspace trust code into ExtensionManifestPropertiesService
