+++ src/vs/workbench/contrib/extensions/browser/extensionsList.ts
@@ -26,7 +26,7 @@ import { foreground, listActiveSelectionForeground, listActiveSelectionBackgroun
 import { WORKBENCH_BACKGROUND } from 'vs/workbench/common/theme';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { localize } from 'vs/nls';
-import { IExtensionWorkspaceTrustRequestService } from 'vs/workbench/services/extensions/common/extensionWorkspaceTrustRequest';
+import { IExtensionManifestPropertiesService } from 'vs/workbench/services/extensions/common/extensionManifestPropertiesService';
 
 export const EXTENSION_LIST_ELEMENT_HEIGHT = 62;
 
@@ -67,7 +67,7 @@ export class Renderer implements IPagedRenderer<IExtension, ITemplateData> {
 		@IExtensionService private readonly extensionService: IExtensionService,
 		@IExtensionManagementServerService private readonly extensionManagementServerService: IExtensionManagementServerService,
 		@IExtensionsWorkbenchService private readonly extensionsWorkbenchService: IExtensionsWorkbenchService,
-		@IExtensionWorkspaceTrustRequestService private readonly extensionWorkspaceTrustRequestService: IExtensionWorkspaceTrustRequestService,
+		@IExtensionManifestPropertiesService private readonly extensionManifestPropertiesService: IExtensionManifestPropertiesService,
 		@IContextMenuService private readonly contextMenuService: IContextMenuService,
 	) { }
 
@@ -207,7 +207,7 @@ export class Renderer implements IPagedRenderer<IExtension, ITemplateData> {
 
 		if (extension.local?.manifest.workspaceTrust?.request) {
 			const trustRequirement = extension.local.manifest.workspaceTrust;
-			const requestType = this.extensionWorkspaceTrustRequestService.getExtensionWorkspaceTrustRequestType(extension.local.manifest);
+			const requestType = this.extensionManifestPropertiesService.getExtensionWorkspaceTrustRequestType(extension.local.manifest);
 			if (requestType !== 'never' && trustRequirement.request !== 'never') {
 				data.workspaceTrustDescription.textContent = trustRequirement.description;
 			} else if (requestType === 'onStart') {
