+++ src/vs/workbench/services/extensions/common/extensionManifestPropertiesService.ts
@@ -4,13 +4,16 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IExtensionManifest, ExtensionKind, ExtensionIdentifier } from 'vs/platform/extensions/common/extensions';
+import { IExtensionManifest, ExtensionKind, ExtensionIdentifier, ExtensionWorkspaceTrustRequestType } from 'vs/platform/extensions/common/extensions';
 import { ExtensionsRegistry } from 'vs/workbench/services/extensions/common/extensionsRegistry';
 import { getGalleryExtensionId } from 'vs/platform/extensionManagement/common/extensionManagementUtil';
 import { isNonEmptyArray } from 'vs/base/common/arrays';
 import { IProductService } from 'vs/platform/product/common/productService';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
+import { ExtensionWorkspaceTrustRequest } from 'vs/base/common/product';
+import { Disposable } from 'vs/base/common/lifecycle';
+import { isWorkspaceTrustEnabled, WORKSPACE_TRUST_EXTENSION_REQUEST } from 'vs/workbench/services/workspaces/common/workspaceTrust';
 
 export const IExtensionManifestPropertiesService = createDecorator<IExtensionManifestPropertiesService>('extensionManifestPropertiesService');
 
@@ -27,9 +30,10 @@ export interface IExtensionManifestPropertiesService {
 
 	getExtensionKind(manifest: IExtensionManifest): ExtensionKind[];
 	canSupportVirtualWorkspace(manifest: IExtensionManifest): boolean;
+	getExtensionWorkspaceTrustRequestType(manifest: IExtensionManifest): ExtensionWorkspaceTrustRequestType;
 }
 
-export class ExtensionManifestPropertiesService implements IExtensionManifestPropertiesService {
+export class ExtensionManifestPropertiesService extends Disposable implements IExtensionManifestPropertiesService {
 
 	readonly _serviceBrand: undefined;
 
@@ -40,10 +44,30 @@ export class ExtensionManifestPropertiesService implements IExtensionManifestPro
 	private _productVirtualWorkspaceSupportMap: Map<string, { default?: boolean, override?: boolean }> | null = null;
 	private _configuredVirtualWorkspaceSupportMap: Map<string, boolean> | null = null;
 
+	private readonly _configuredExtensionWorkspaceTrustRequestMap: Map<string, { request: ExtensionWorkspaceTrustRequestType, version?: string }>;
+	private readonly _productExtensionWorkspaceTrustRequestMap: Map<string, ExtensionWorkspaceTrustRequest>;
+
 	constructor(
 		@IProductService private readonly productService: IProductService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
-	) { }
+	) {
+		super();
+
+		// Workspace trust request type (settings.json)
+		this._configuredExtensionWorkspaceTrustRequestMap = new Map<string, { request: ExtensionWorkspaceTrustRequestType, version?: string }>();
+		const configuredExtensionWorkspaceTrustRequests = configurationService.inspect<{ [key: string]: { request: ExtensionWorkspaceTrustRequestType, version?: string } }>(WORKSPACE_TRUST_EXTENSION_REQUEST).userValue || {};
+		for (const id of Object.keys(configuredExtensionWorkspaceTrustRequests)) {
+			this._configuredExtensionWorkspaceTrustRequestMap.set(ExtensionIdentifier.toKey(id), configuredExtensionWorkspaceTrustRequests[id]);
+		}
+
+		// Workpace trust request type (products.json)
+		this._productExtensionWorkspaceTrustRequestMap = new Map<string, ExtensionWorkspaceTrustRequest>();
+		if (productService.extensionWorkspaceTrustRequest) {
+			for (const id of Object.keys(productService.extensionWorkspaceTrustRequest)) {
+				this._productExtensionWorkspaceTrustRequestMap.set(ExtensionIdentifier.toKey(id), productService.extensionWorkspaceTrustRequest[id]);
+			}
+		}
+	}
 
 	prefersExecuteOnUI(manifest: IExtensionManifest): boolean {
 		const extensionKind = this.getExtensionKind(manifest);
@@ -123,6 +147,40 @@ export class ExtensionManifestPropertiesService implements IExtensionManifestPro
 
 		// Default - supports virtual workspace
 		return true;
+	}
+	getExtensionWorkspaceTrustRequestType(manifest: IExtensionManifest): ExtensionWorkspaceTrustRequestType {
+		// Workspace trust feature is disabled, or extension has no entry point
+		if (!isWorkspaceTrustEnabled(this.configurationService) || !manifest.main) {
+			return 'never';
+		}
+
+		// Get extension workspace trust requirements from settings.json
+		const configuredWorkspaceTrustRequest = this.getConfiguredExtensionWorkspaceTrustRequest(manifest);
+
+		// Get extension workspace trust requirements from product.json
+		const productWorkspaceTrustRequest = this.getProductExtensionWorkspaceTrustRequest(manifest);
+
+		// Use settings.json override value if it exists
+		if (configuredWorkspaceTrustRequest) {
+			return configuredWorkspaceTrustRequest;
+		}
+
+		// Use product.json override value if it exists
+		if (productWorkspaceTrustRequest?.override) {
+			return productWorkspaceTrustRequest.override;
+		}
+
+		// Use extension manifest value if it exists
+		if (manifest.workspaceTrust?.request !== undefined) {
+			return manifest.workspaceTrust.request;
+		}
+
+		// Use product.json default value if it exists
+		if (productWorkspaceTrustRequest?.default) {
+			return productWorkspaceTrustRequest.default;
+		}
+
+		return 'onStart';
 	}
 
 	deduceExtensionKind(manifest: IExtensionManifest): ExtensionKind[] {
@@ -224,6 +282,22 @@ export class ExtensionManifestPropertiesService implements IExtensionManifestPro
 
 		const extensionId = getGalleryExtensionId(manifest.publisher, manifest.name);
 		return this._configuredVirtualWorkspaceSupportMap.get(ExtensionIdentifier.toKey(extensionId));
+	}
+	
+	private getConfiguredExtensionWorkspaceTrustRequest(manifest: IExtensionManifest): ExtensionWorkspaceTrustRequestType | undefined {
+		const extensionId = getGalleryExtensionId(manifest.publisher, manifest.name);
+		const extensionWorkspaceTrustRequest = this._configuredExtensionWorkspaceTrustRequestMap.get(ExtensionIdentifier.toKey(extensionId));
+
+		if (extensionWorkspaceTrustRequest && (extensionWorkspaceTrustRequest.version === undefined || extensionWorkspaceTrustRequest.version === manifest.version)) {
+			return extensionWorkspaceTrustRequest.request;
+		}
+
+		return undefined;
+	}
+
+	private getProductExtensionWorkspaceTrustRequest(manifest: IExtensionManifest): ExtensionWorkspaceTrustRequest | undefined {
+		const extensionId = getGalleryExtensionId(manifest.publisher, manifest.name);
+		return this._productExtensionWorkspaceTrustRequestMap.get(ExtensionIdentifier.toKey(extensionId));
 	}
 
 	private toArray(extensionKind: ExtensionKind | ExtensionKind[]): ExtensionKind[] {
